+++ rtgui/mycurve.cc
@@ -175,7 +175,7 @@ bool MyCurve::snapCoordinateY(double testedVal, double realVal)
 
 float MyCurve::getVal(LUTf &curve, int x)
 {
-    if ((graphW - 2) == curve.getSize()) {
+    if (size_t(graphW - 2) == curve.getSize()) {
         return curve[x];
     } else {
         return curve.getVal01(float(x) / (graphW - 3));
